refactor(mcp): zero Thermos quality findings
