+++ scripts/gen_mcp_server_json.py
@@ -24,7 +24,7 @@ from typing import TYPE_CHECKING, Any
 
 import jsonschema
 
-from mergecraft.cli.mcp_serve import build_codegen_tool_context
+from mergecraft.cli.mcp_serve import codegen_tool_context
 from mergecraft.mcp.public import PUBLIC_TOOL_NAMES, build_public_tools
 from mergecraft.mcp.rpc import package_version
 
@@ -42,8 +42,8 @@ GENERATED_HEADER = "<!-- Generated by scripts/gen_mcp_server_json.py; do not edi
 
 
 def _public_tool_specs() -> list[ToolSpec]:
-    ctx = build_codegen_tool_context()
-    specs = build_public_tools(ctx)
+    with codegen_tool_context() as ctx:
+        specs = build_public_tools(ctx)
     names = {spec.name for spec in specs}
     if names != set(PUBLIC_TOOL_NAMES):
         msg = f"public tool names mismatch: {names} != {set(PUBLIC_TOOL_NAMES)}"
