+++ src/mergecraft/cli/explain_cmd.py
@@ -22,6 +22,7 @@ from mergecraft.review.completed import (
     completed_review_exists,
     lookup_finding_packet_in_review,
 )
+from mergecraft.review.explain import finding_explain_payload
 
 
 def _read_diff(repo_root: Path) -> str:
@@ -55,27 +56,6 @@ def _change_payload(repo_root: Path) -> dict[str, Any]:
             else "No working-tree diff to explain."
         ),
     }
-
-
-def finding_explain_payload(
-    finding_id: str,
-    packet: dict[str, Any],
-    *,
-    review_id: str | None = None,
-) -> dict[str, Any]:
-    state = packet.get("state", "unverified")
-    kinds = packet.get("kinds", [])
-    kinds_text = ", ".join(str(item) for item in kinds) if isinstance(kinds, list) else "none"
-    payload: dict[str, Any] = {
-        "verb": "explain",
-        "finding_id": finding_id,
-        "paths": [],
-        "summary": f"Finding {finding_id} is {state} (kinds: {kinds_text}).",
-        "packet": packet,
-    }
-    if review_id is not None:
-        payload["review_id"] = review_id
-    return payload
 
 
 def _render_table(payload: dict[str, Any]) -> Table:
